requirements.txt
colorama==0.4.6
peewee==3.17.8
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
